requirements.txt
Django==4.2.7
stripe==7.4.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
whitenoise==6.6.0
